requirements.txt
streamlit
agno
openai
python-dotenv
streamlit-chat
google-genai
anthropic
openai
googlesearch-python
sqlalchemy
duckduckgo-search
